Expanded Unit Handler to include more information
Shield burst, antimatter, strike-craft values, regeneration

+++ src/main/java/org/dshaver/sins/domain/ingest/unit/AdditiveValues.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class AdditiveValues {
     double maxAntimatter;
     double antimatterRestoreRate;
-    double maxSquadronCapacity;
+    int maxSquadronCapacity;
 
     
 }
